+++ components/ProjectShowcase/ProjectShowcase.tsx
@@ -2,7 +2,7 @@
 
 import Image from "next/legacy/image";
 
-import { Project } from "../../app/(pages)/work/page";
+import { Project } from "../../app/[locale]/work/page";
 import { Link } from "next-intl";
 import { RichText } from "@graphcms/rich-text-react-renderer";
 import { Button } from "../Button";
